requirements.txt
# install requirements with: pip install -r requirements.txt

streamlit
Pillow
albumentations
opencv-python
pyyaml
numpy
